[FIX] Prevent clean action if the file does not exist or the variable is empty. Skip on option 2.

+++ flux_common.sh
@@ -1905,19 +1905,19 @@ function bootstrap_new() {
 	if [[ -f "$FILE_PATH" ]]; then
 		if [[ "$Mode" != "install" ]]; then
 			start_service
-			# if whiptail --yesno "Would you like remove bootstrap archive file?" 8 60; then
+			if [[ -f $FILE_PATH && $FILE_PATH != "" ]]; then
 				sudo rm -rf $FILE_PATH > /dev/null 2>&1 && sleep 2
-			# fi
+			fi
 		fi
 		return
 	else
 		if [[ ! -f $DATA_PATH/install_conf.json ]]; then
 			bootstrap_manual
-			if [[ "$Mode" != "install" && "$server_offline" == "0" ]]; then
+			if [[ "$Mode" != "install" && "$server_offline" == "0" && CHOICE == "1)" ]]; then
 				start_service
-				# if whiptail --yesno "Would you like remove bootstrap archive file?" 8 60; then
-				  sudo rm -rf $DATA_PATH/$BOOTSTRAP_FILE > /dev/null 2>&1 && sleep 2
-				# fi
+        if [[ -f $DATA_PATH/$BOOTSTRAP_FILE && $BOOTSTRAP_FILE != "" ]]; then
+				    sudo rm -rf $DATA_PATH/$BOOTSTRAP_FILE > /dev/null 2>&1 && sleep 2
+        fi
 			fi
 			return
 		fi
